requirements.txt
discord.py
python-dotenv
psutil
aiofiles
asyncio-throttle
